added some routes

+++ src/controllers/post.controller.ts
@@ -2,7 +2,7 @@ import { response, type NextFunction, type Request, type Response } from "expres
 import { asyncHandler } from "../utils/AsyncHandler.js";
 import { PostSchema, PostUpdateSchema } from "../validators/post.validator.js";
 import { ApiError } from "../utils/ApiError.js";
-import { createPost, deletePost, getPostByCategories, getPostById, updatePost } from "../services/post.service.js";
+import { createPost, deletePost, getAllPosts, getPostByCategories, getPostById, updatePost } from "../services/post.service.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 
 export const create = asyncHandler(async(req:Request, res:Response, _:NextFunction)=>{
@@ -28,6 +28,33 @@ export const update = asyncHandler(async(req:Request, res:Response, _:NextFuncti
     if(!success){
         throw new ApiError(402, "Invalid feilds");
     }
+    const post = await getPostById(parseInt(postId || ""));
+    if(!post){
+        throw new ApiError(404, "Post not found");
+    }
+    const updatedPost = await updatePost({content, title, id: parseInt(postId || "")});
+    if(!updatedPost){
+        throw new ApiError(405, "Post cannot be updated");
+    }
+    return res.status(200).json(new ApiResponse(200, {post:updatedPost}, "Post updated successfully"));
+});
+
+export const updateOwn = asyncHandler(async(req:Request, res:Response, _:NextFunction)=>{
+    const {content, title} = req.body;
+    const userId = req.userId;
+    const {success} = PostUpdateSchema.safeParse({title, content});
+    const postId = req.params.id;
+    
+    if(!success){
+        throw new ApiError(402, "Invalid feilds");
+    }
+    const post = await getPostById(parseInt(postId || ""));
+    if(!post){
+        throw new ApiError(404, "Post not found");
+    }
+    if(post.userId!=userId){
+        throw new ApiError(403, "You are not the owner of the post");
+    }
     const updatedPost = await updatePost({content, title, id: parseInt(postId || "")});
     if(!updatedPost){
         throw new ApiError(405, "Post cannot be updated");
@@ -37,6 +64,22 @@ export const update = asyncHandler(async(req:Request, res:Response, _:NextFuncti
 
 export const deleteAPost = asyncHandler(async(req:Request, res:Response, _:NextFunction)=>{
     const postId = req.params.id;
+    const isDeleted = await deletePost(parseInt(postId || ""));
+    if(!isDeleted){
+        throw new ApiError(406, "Post cannot be deleted");
+    }
+    return res.status(200).json(new ApiResponse(200, {success:true}, "Post deleted successfully"));
+});
+
+export const deleteOwn = asyncHandler(async(req:Request, res:Response, _:NextFunction)=>{
+    const postId = req.params.id;
+    const post = await getPostById(parseInt(postId || ""));
+    if(!post){
+        throw new ApiError(404, "Post not found");
+    }
+    if(post.userId!=req.userId){
+        throw new ApiError(403, "You are not the owner");
+    }
     const isDeleted = await deletePost(parseInt(postId || ""));
     if(!isDeleted){
         throw new ApiError(406, "Post cannot be deleted");
@@ -59,5 +102,10 @@ export const getPostsByCategoryId = asyncHandler(async(req:Request, res:Response
     if(!posts){
         throw new ApiError(404, "Posts not found");
     }
+    return res.status(200).json(new ApiResponse(200, {posts}, "Posts sent successfully"));
+});
+
+export const getPostsAll = asyncHandler(async(req:Request, res:Response, _:NextFunction)=>{
+    const posts = await getAllPosts();
     return res.status(200).json(new ApiResponse(200, {posts}, "Posts sent successfully"));
 });
